move forecast horizon to read forecast values
